+++ x-pack/plugins/maps/common/descriptor_types/map_descriptor.ts
@@ -11,7 +11,7 @@ import { ReactNode } from 'react';
 import { GeoJsonProperties } from 'geojson';
 import { Geometry } from 'geojson';
 import { Query } from '../../../../../src/plugins/data/common';
-import { DRAW_TYPE, ES_GEO_FIELD_TYPE, ES_SPATIAL_RELATIONS } from '../constants';
+import { DRAW_TYPE, ES_SPATIAL_RELATIONS } from '../constants';
 
 export type MapExtent = {
   minLon: number;
@@ -70,9 +70,6 @@ export type DrawState = {
   actionId: string;
   drawType: DRAW_TYPE;
   filterLabel?: string; // point radius filter alias
-  geoFieldName?: string;
-  geoFieldType?: ES_GEO_FIELD_TYPE;
   geometryLabel?: string;
-  indexPatternId?: string;
   relation?: ES_SPATIAL_RELATIONS;
 };
