+++ x-pack/plugins/maps/common/elasticsearch_util/index.ts
@@ -8,4 +8,6 @@
 export * from './es_agg_utils';
 export * from './convert_to_geojson';
 export * from './elasticsearch_geo_utils';
+export * from './spatial_filter_utils';
+export * from './types';
 export { isTotalHitsGreaterThan, TotalHits } from './total_hits';
