+++ x-pack/plugins/maps/public/components/distance_filter_form.tsx
@@ -16,45 +16,26 @@ import {
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { ActionExecutionContext, Action } from 'src/plugins/ui_actions/public';
-import { MultiIndexGeoFieldSelect } from './multi_index_geo_field_select';
-import { GeoFieldWithIndex } from './geo_field_with_index';
 import { ActionSelect } from './action_select';
 import { ACTION_GLOBAL_APPLY_FILTER } from '../../../../../src/plugins/data/public';
 
 interface Props {
   className?: string;
   buttonLabel: string;
-  geoFields: GeoFieldWithIndex[];
   getFilterActions?: () => Promise<Action[]>;
   getActionContext?: () => ActionExecutionContext;
-  onSubmit: ({
-    actionId,
-    filterLabel,
-    indexPatternId,
-    geoFieldName,
-  }: {
-    actionId: string;
-    filterLabel: string;
-    indexPatternId: string;
-    geoFieldName: string;
-  }) => void;
+  onSubmit: ({ actionId, filterLabel }: { actionId: string; filterLabel: string }) => void;
 }
 
 interface State {
   actionId: string;
-  selectedField: GeoFieldWithIndex | undefined;
   filterLabel: string;
 }
 
 export class DistanceFilterForm extends Component<Props, State> {
   state: State = {
     actionId: ACTION_GLOBAL_APPLY_FILTER,
-    selectedField: this.props.geoFields.length ? this.props.geoFields[0] : undefined,
     filterLabel: '',
-  };
-
-  _onGeoFieldChange = (selectedField: GeoFieldWithIndex | undefined) => {
-    this.setState({ selectedField });
   };
 
   _onFilterLabelChange = (e: ChangeEvent<HTMLInputElement>) => {
@@ -68,14 +49,9 @@ export class DistanceFilterForm extends Component<Props, State> {
   };
 
   _onSubmit = () => {
-    if (!this.state.selectedField) {
-      return;
-    }
     this.props.onSubmit({
       actionId: this.state.actionId,
       filterLabel: this.state.filterLabel,
-      indexPatternId: this.state.selectedField.indexPatternId,
-      geoFieldName: this.state.selectedField.geoFieldName,
     });
   };
 
@@ -95,12 +71,6 @@ export class DistanceFilterForm extends Component<Props, State> {
           />
         </EuiFormRow>
 
-        <MultiIndexGeoFieldSelect
-          selectedField={this.state.selectedField}
-          fields={this.props.geoFields}
-          onChange={this._onGeoFieldChange}
-        />
-
         <ActionSelect
           getFilterActions={this.props.getFilterActions}
           getActionContext={this.props.getActionContext}
@@ -111,7 +81,7 @@ export class DistanceFilterForm extends Component<Props, State> {
         <EuiSpacer size="m" />
 
         <EuiTextAlign textAlign="right">
-          <EuiButton size="s" fill onClick={this._onSubmit} isDisabled={!this.state.selectedField}>
+          <EuiButton size="s" fill onClick={this._onSubmit}>
             {this.props.buttonLabel}
           </EuiButton>
         </EuiTextAlign>
