+++ x-pack/plugins/maps/public/connected_components/map_container/map_container.tsx
@@ -20,15 +20,12 @@ import { ToolbarOverlay } from '../toolbar_overlay';
 import { EditLayerPanel } from '../edit_layer_panel';
 import { AddLayerPanel } from '../add_layer_panel';
 import { ExitFullScreenButton } from '../../../../../../src/plugins/kibana_react/public';
-import { getIndexPatternsFromIds } from '../../index_pattern_util';
-import { ES_GEO_FIELD_TYPE, RawValue } from '../../../common/constants';
-import { indexPatterns as indexPatternsUtils } from '../../../../../../src/plugins/data/public';
+import { RawValue } from '../../../common/constants';
 import { FLYOUT_STATE } from '../../reducers/ui';
 import { MapSettings } from '../../reducers/map';
 import { MapSettingsPanel } from '../map_settings_panel';
 import { registerLayerWizards } from '../../classes/layers/load_layer_wizards';
 import { RenderToolTipContent } from '../../classes/tooltips/tooltip_property';
-import { GeoFieldWithIndex } from '../../components/geo_field_with_index';
 import { MapRefreshConfig } from '../../../common/descriptor_types';
 import { ILayer } from '../../classes/layers/layer';
 
@@ -58,7 +55,6 @@ export interface Props {
 interface State {
   isInitialLoadRenderTimeoutComplete: boolean;
   domId: string;
-  geoFields: GeoFieldWithIndex[];
   showFitToBoundsButton: boolean;
   showTimesliderButton: boolean;
 }
@@ -66,7 +62,6 @@ interface State {
 export class MapContainer extends Component<Props, State> {
   private _isMounted: boolean = false;
   private _isInitalLoadRenderTimerStarted: boolean = false;
-  private _prevIndexPatternIds: string[] = [];
   private _refreshTimerId: number | null = null;
   private _prevIsPaused: boolean | null = null;
   private _prevInterval: number | null = null;
@@ -74,7 +69,6 @@ export class MapContainer extends Component<Props, State> {
   state: State = {
     isInitialLoadRenderTimeoutComplete: false,
     domId: uuid(),
-    geoFields: [],
     showFitToBoundsButton: false,
     showTimesliderButton: false,
   };
@@ -94,10 +88,6 @@ export class MapContainer extends Component<Props, State> {
     if (this.props.areLayersLoaded && !this._isInitalLoadRenderTimerStarted) {
       this._isInitalLoadRenderTimerStarted = true;
       this._startInitialLoadRenderTimer();
-    }
-
-    if (!!this.props.addFilters) {
-      this._loadGeoFields(this.props.indexPatternIds);
     }
   }
 
@@ -149,40 +139,6 @@ export class MapContainer extends Component<Props, State> {
     if (this._isMounted && this.state.showTimesliderButton !== showTimesliderButton) {
       this.setState({ showTimesliderButton });
     }
-  }
-
-  async _loadGeoFields(nextIndexPatternIds: string[]) {
-    if (_.isEqual(nextIndexPatternIds, this._prevIndexPatternIds)) {
-      // all ready loaded index pattern ids
-      return;
-    }
-
-    this._prevIndexPatternIds = nextIndexPatternIds;
-
-    const geoFields: GeoFieldWithIndex[] = [];
-    const indexPatterns = await getIndexPatternsFromIds(nextIndexPatternIds);
-    indexPatterns.forEach((indexPattern) => {
-      indexPattern.fields.forEach((field) => {
-        if (
-          indexPattern.id &&
-          !indexPatternsUtils.isNestedField(field) &&
-          (field.type === ES_GEO_FIELD_TYPE.GEO_POINT || field.type === ES_GEO_FIELD_TYPE.GEO_SHAPE)
-        ) {
-          geoFields.push({
-            geoFieldName: field.name,
-            geoFieldType: field.type,
-            indexPatternTitle: indexPattern.title,
-            indexPatternId: indexPattern.id,
-          });
-        }
-      });
-    });
-
-    if (!this._isMounted) {
-      return;
-    }
-
-    this.setState({ geoFields });
   }
 
   _setRefreshTimer = () => {
@@ -289,13 +245,11 @@ export class MapContainer extends Component<Props, State> {
             getFilterActions={getFilterActions}
             getActionContext={getActionContext}
             onSingleValueTrigger={onSingleValueTrigger}
-            geoFields={this.state.geoFields}
             renderTooltipContent={renderTooltipContent}
           />
           {!this.props.settings.hideToolbarOverlay && (
             <ToolbarOverlay
               addFilters={addFilters}
-              geoFields={this.state.geoFields}
               getFilterActions={getFilterActions}
               getActionContext={getActionContext}
               showFitToBoundsButton={this.state.showFitToBoundsButton}
