+++ x-pack/plugins/maps/public/connected_components/mb_map/draw_control/draw_filter_control/index.ts
@@ -10,13 +10,18 @@ import { ThunkDispatch } from 'redux-thunk';
 import { connect } from 'react-redux';
 import { DrawFilterControl } from './draw_filter_control';
 import { updateDrawState } from '../../../../actions';
-import { getDrawState, isDrawingFilter } from '../../../../selectors/map_selectors';
+import {
+  getDrawState,
+  isDrawingFilter,
+  getGeoFieldNames,
+} from '../../../../selectors/map_selectors';
 import { MapStoreState } from '../../../../reducers/store';
 
 function mapStateToProps(state: MapStoreState) {
   return {
     isDrawingFilter: isDrawingFilter(state),
     drawState: getDrawState(state),
+    geoFieldNames: getGeoFieldNames(state),
   };
 }
