+++ x-pack/plugins/maps/public/connected_components/mb_map/tooltip_control/tooltip_control.test.tsx
@@ -77,7 +77,7 @@ const defaultProps = {
   layerList: [mockLayer],
   isDrawingFilter: false,
   addFilters: async () => {},
-  geoFields: [],
+  geoFieldNames: [],
   openTooltips: [],
   hasLockedTooltips: false,
 };
