+++ x-pack/plugins/translations/translations/ja-JP.json
@@ -13312,7 +13312,6 @@
     "xpack.maps.emsSource.tooltipsTitle": "ツールチップフィールド",
     "xpack.maps.es_geo_utils.convert.invalidGeometryCollectionErrorMessage": "GeometryCollectionを convertESShapeToGeojsonGeometryに渡さないでください",
     "xpack.maps.es_geo_utils.convert.unsupportedGeometryTypeErrorMessage": "{geometryType} ジオメトリから Geojson に変換できません。サポートされていません",
-    "xpack.maps.es_geo_utils.distanceFilterAlias": "{pointLabel} の {distanceKm}km 以内にある {geoFieldName}",
     "xpack.maps.es_geo_utils.unsupportedFieldTypeErrorMessage": "サポートされていないフィールドタイプ、期待値：{expectedTypes}、提供された値：{fieldType}",
     "xpack.maps.es_geo_utils.unsupportedGeometryTypeErrorMessage": "サポートされていないジオメトリタイプ、期待値：{expectedTypes}、提供された値：{geometryType}",
     "xpack.maps.es_geo_utils.wkt.invalidWKTErrorMessage": "{wkt} を Geojson に変換できません。有効な WKT が必要です。",
@@ -13492,7 +13491,6 @@
     "xpack.maps.metricSelect.selectAggregationPlaceholder": "集約を選択",
     "xpack.maps.metricSelect.sumDropDownOptionLabel": "合計",
     "xpack.maps.metricSelect.termsDropDownOptionLabel": "トップ用語",
-    "xpack.maps.multiIndexFieldSelect.fieldLabel": "フィールドのフィルタリング",
     "xpack.maps.mvtSource.addFieldLabel": "追加",
     "xpack.maps.mvtSource.fieldPlaceholderText": "フィールド名",
     "xpack.maps.mvtSource.numberFieldLabel": "数字",
