+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -13488,7 +13488,6 @@
     "xpack.maps.emsSource.tooltipsTitle": "工具提示字段",
     "xpack.maps.es_geo_utils.convert.invalidGeometryCollectionErrorMessage": "不应将 GeometryCollection 传递给 convertESShapeToGeojsonGeometry",
     "xpack.maps.es_geo_utils.convert.unsupportedGeometryTypeErrorMessage": "无法将 {geometryType} 几何图形转换成 geojson，不支持",
-    "xpack.maps.es_geo_utils.distanceFilterAlias": "{pointLabel} {distanceKm}km 内的 {geoFieldName}",
     "xpack.maps.es_geo_utils.unsupportedFieldTypeErrorMessage": "字段类型不受支持，应为 {expectedTypes}，而提供的是 {fieldType}",
     "xpack.maps.es_geo_utils.unsupportedGeometryTypeErrorMessage": "几何类型不受支持，应为 {expectedTypes}，而提供的是 {geometryType}",
     "xpack.maps.es_geo_utils.wkt.invalidWKTErrorMessage": "无法将 {wkt} 转换成 geojson。需要有效的 WKT。",
@@ -13669,7 +13668,6 @@
     "xpack.maps.metricSelect.selectAggregationPlaceholder": "选择聚合",
     "xpack.maps.metricSelect.sumDropDownOptionLabel": "求和",
     "xpack.maps.metricSelect.termsDropDownOptionLabel": "热门词",
-    "xpack.maps.multiIndexFieldSelect.fieldLabel": "筛选字段",
     "xpack.maps.mvtSource.addFieldLabel": "添加",
     "xpack.maps.mvtSource.fieldPlaceholderText": "字段名称",
     "xpack.maps.mvtSource.numberFieldLabel": "数字",
